Fix temporal files for the processes in windows

+++ core/arrnorm.py
@@ -20,6 +20,7 @@
 """
 import os
 import platform
+import shutil
 import sys
 import subprocess
 
@@ -60,6 +61,10 @@ class Normalization:
         self.output = output
         self.feedback = feedback
 
+        self.img_norm = None
+        self.no_neg = None
+        self.norm_masked = None
+
     def run(self):
 
         self.feedback.pushInfo("PROCESSING IMAGE: {target}".format(target=os.path.basename(self.img_target)))
@@ -86,7 +91,15 @@ class Normalization:
         if self.nodata_mask:
             self.feedback.setProgress(90)
             self.make_mask()
-            self.apply_mask()
+            self.apply_mask(self.no_neg if self.no_neg else self.img_norm)
+
+        # finish
+        if self.norm_masked:
+            shutil.move(self.norm_masked, self.output)
+        elif self.no_neg:
+            shutil.move(self.no_neg, self.output)
+        else:
+            shutil.move(self.img_norm, self.output)
 
         self.clean()
 
@@ -107,6 +120,7 @@ class Normalization:
         pixel_size_x = abs(round(gdal_file.GetGeoTransform()[1], 3))
         pixel_size_y = abs(round(gdal_file.GetGeoTransform()[5], 3))
         if pixel_size_x != pixel_size_y:
+            self.clean()
             raise QgsProcessingException('\nError: target and reference image have different pixel sizes\n')
 
         # extent
@@ -136,6 +150,7 @@ class Normalization:
         if cmd_out.returncode == 0:  # successfully
             self.feedback.pushInfo('Clipped ref image successfully: ' + os.path.basename(self.img_ref_clip))
         else:
+            self.clean()
             raise QgsProcessingException('\nError clipping reference image: ' + str(cmd_out.stderr.decode()))
 
     def imad(self):
@@ -152,24 +167,31 @@ class Normalization:
         # ======================================
         # Radcal process
 
+        filename, ext = os.path.splitext(os.path.basename(self.img_target))
+        self.img_norm = os.path.join(os.path.dirname(os.path.abspath(self.img_target)), filename + "_radcal" + ext)
+
         self.feedback.pushInfo("\nRadcal process for\n" +
               os.path.basename(self.img_ref) + " " + os.path.basename(self.img_target) + " " +
               " with iMad image: " + os.path.basename(self.img_imad))
-        self.img_norm = radcal.main(self.img_imad, self.output, ncpThresh=self.prob_thres)
+        radcal.main(self.img_imad, self.img_norm, ncpThresh=self.prob_thres)
 
     def no_negative_value(self, image):
         # ======================================
         # Convert negative values to NoData to image normalized
 
+        filename, ext = os.path.splitext(os.path.basename(self.img_target))
+        self.no_neg = os.path.join(os.path.dirname(os.path.abspath(self.img_target)), filename + "_no_neg" + ext)
+
         self.feedback.pushInfo('\nConverting negative values for\n' +
               os.path.basename(os.path.basename(image)))
         cmd = ['gdal_calc' if platform.system() == 'Windows' else 'gdal_calc.py', '-A', '"{}"'.format(image),
                '--overwrite', '--calc', '"A*(A>=0)"', '--allBands=A', '--NoDataValue=0', '--type=UInt16', '--quiet',
-               '--outfile', '"{}"'.format(image)]
+               '--outfile', '"{}"'.format(self.no_neg)]
         cmd_out = subprocess.run(" ".join(cmd), shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         if cmd_out.returncode == 0:  # successfully
             self.feedback.pushInfo('Negative values converted successfully: ' + os.path.basename(image))
         else:
+            self.clean()
             raise QgsProcessingException('\nError converting values: ' + str(cmd_out.stderr.decode()))
 
     def make_mask(self):
@@ -190,27 +212,35 @@ class Normalization:
         if cmd_out.returncode == 0:  # successfully
             self.feedback.pushInfo('Mask created successfully: ' + os.path.basename(self.mask_file))
         else:
+            self.clean()
             raise QgsProcessingException('\nError creating mask: ' + str(cmd_out.stderr.decode()))
 
-    def apply_mask(self):
+    def apply_mask(self, image):
         # ======================================
         # Apply mask to image normalized
 
+        filename, ext = os.path.splitext(os.path.basename(self.img_target))
+        self.norm_masked = os.path.join(os.path.dirname(os.path.abspath(self.img_target)), filename + "_norm_masked" + ext)
+
         self.feedback.pushInfo('\nApplying mask for\n' +
-              os.path.basename(self.img_target) + " " + os.path.basename(self.img_norm))
-        cmd = 'gdal_calc' if platform.system() == 'Windows' else 'gdal_calc.py' + ' -A "' + self.img_norm + '" -B "' + \
+              os.path.basename(self.img_target) + " " + os.path.basename(image))
+        cmd = 'gdal_calc' if platform.system() == 'Windows' else 'gdal_calc.py' + ' -A "' + image + '" -B "' + \
               self.mask_file + '" --type=UInt16 --co COMPRESS=LZW --co PREDICTOR=2 TILED=YES --outfile="' + \
-              self.img_norm + '" --calc="A*(B==1)" --NoDataValue=0  --allBands=A  --overwrite --quiet'
+              self.norm_masked + '" --calc="A*(B==1)" --NoDataValue=0  --allBands=A  --overwrite --quiet'
         cmd_out = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         if cmd_out.returncode == 0:  # successfully
             self.feedback.pushInfo('Mask applied successfully: ' + os.path.basename(self.mask_file))
         else:
+            self.clean()
             raise QgsProcessingException('\nError applied mask: ' + str(cmd_out.stderr.decode()))
 
     def clean(self):
         # delete the MAD file
-        os.remove(self.img_imad)
+        os.remove(self.img_imad) if os.path.exists(self.img_imad) else None
         # delete the clip reference image
         if self.img_ref_clip != self.img_ref:
-            os.remove(self.img_ref_clip)
+            os.remove(self.img_ref_clip) if os.path.exists(self.img_ref_clip) else None
+        os.remove(self.img_norm) if self.img_norm and os.path.exists(self.img_norm) else None
+        os.remove(self.no_neg) if self.no_neg and os.path.exists(self.no_neg) else None
+        os.remove(self.norm_masked) if self.norm_masked and os.path.exists(self.norm_masked) else None
